+++ sample/src/main.ts
@@ -1,0 +1,5 @@
+import { mainScene } from "./mainScene";
+
+module.exports = () => {
+    g.game.pushScene(mainScene());
+};