+++ sample/src/mainScene2.ts
@@ -1,32 +1,27 @@
-import {Label, RubyAlign} from "@akashic-extension/akashic-label";
-var game = g.game;
+import { Label, RubyAlign } from "@akashic-extension/akashic-label";
+import { mainScene3 } from "./mainScene3";
 
-module.exports = function() {
+export function mainScene2(): g.Scene {
+	var game = g.game;
 	var scene = new g.Scene({
 		game: game,
 		assetIds: ["bmpfont", "bmpfont-glyph", "mplus", "mplus-glyph"]
 	});
 	var rate = game.fps / 6;
-	scene.loaded.add(function() {
+	scene.onLoad.add(() => {
 
 		// グリフデータの生成
-		var mplusGlyph = JSON.parse((<g.TextAsset>scene.assets["mplus-glyph"]).data);
-
+		var mPlusGlyphInfo = JSON.parse(scene.asset.getTextById("mplus-glyph").data);
 		// ビットマップフォント画像とグリフ情報からBitmapFontのインスタンスを生成
 		var mplusfont = new g.BitmapFont({
-			src: scene.assets["mplus"],
-			map: mplusGlyph.map,
-			defaultGlyphWidth: mplusGlyph.width,
-			defaultGlyphHeight: mplusGlyph.height,
-			missingGlyph: mplusGlyph.missingGlyph
+			src: scene.asset.getImageById("mplus"),
+			glyphInfo: mPlusGlyphInfo
 		});
-		var glyph = JSON.parse((<g.TextAsset>scene.assets["bmpfont-glyph"]).data);
+
+		var bmpGlyphInfo = JSON.parse(scene.asset.getTextById("bmpfont-glyph").data);
 		var bmpfont = new g.BitmapFont({
-			src: scene.assets["bmpfont"],
-			map: glyph.map,
-			defaultGlyphWidth: glyph.width,
-			defaultGlyphHeight: glyph.height,
-			missingGlyph: glyph.missingGlyph
+			src: scene.asset.getImageById("bmpfont"),
+			glyphInfo: bmpGlyphInfo
 		});
 
 		var dhint: g.DynamicFontHint = {
@@ -35,10 +30,10 @@ module.exports = function() {
 			maxAtlasWidth: 256,
 			maxAtlasHeight: 256,
 			maxAtlasNum: 8
-		}
+		};
 		var dfont = new g.DynamicFont({
 			game: scene.game,
-			fontFamily: g.FontFamily.Monospace,
+			fontFamily: "monospace",
 			size: 40,
 			hint: dhint
 		});
@@ -50,7 +45,7 @@ module.exports = function() {
 			font: mplusfont,
 			fontSize: 30,
 			width: game.width,
-			textAlign: g.TextAlign.Center
+			textAlign: "center"
 		});
 		tlabel0.x = 0;
 		scene.append(tlabel0);
@@ -60,7 +55,7 @@ module.exports = function() {
 		// ルビの利用
 		var label01 = new Label({
 			scene: scene,
-			text:'use {"rt":"ruby","rb":"ruby"}.',
+			text: `use {"rt":"ruby","rb":"ruby"}.`,
 			font: bmpfont,
 			fontSize: 20,
 			width: 180
@@ -71,7 +66,7 @@ module.exports = function() {
 		// ルビを使わない
 		var label02 = new Label({
 			scene: scene,
-			text:'unuse {"rt":"ruby","rb":"ruby"}.',
+			text: `unuse {"rt":"ruby","rb":"ruby"}.`,
 			font: bmpfont,
 			fontSize: 20,
 			width: game.width,
@@ -84,7 +79,7 @@ module.exports = function() {
 		var counter03 = 0;
 		var label03 = new Label({
 			scene: scene,
-			text:'{"rt":"るび","rb":"ルビ"}の行間',
+			text: `{"rt":"るび","rb":"ルビ"}の行間`,
 			font: mplusfont,
 			fontSize: 20,
 			width: 100,
@@ -93,11 +88,11 @@ module.exports = function() {
 		label03.x = 0;
 		label03.y = y0 + 90;
 		label03.touchable = true;
-		label03.update.add(function() {
+		label03.onUpdate.add(() => {
 			if (game.age % rate === 0) {
-				this.rubyOptions.rubyGap = counter03 % 4 - 5;
+				label03.rubyOptions.rubyGap = counter03 % 4 - 5;
 				counter03++;
-				this.invalidate();
+				label03.invalidate();
 			}
 		}, label03);
 		scene.append(label03);
@@ -106,28 +101,28 @@ module.exports = function() {
 		var counter04 = 0;
 		var label04 = new Label({
 			scene: scene,
-			text:'{"rt":"るび","rb":"ルビ"}サイズ',
+			text: `{"rt":"るび","rb":"ルビ"}サイズ`,
 			font: mplusfont,
 			fontSize: 20,
 			width: game.width,
-			rubyOptions: {rubyFontSize: 15,rubyGap: -5}
+			rubyOptions: {rubyFontSize: 15, rubyGap: -5}
 		});
 		label04.x = 100;
 		label04.y = y0 + 90;
 		label04.touchable = true;
 		scene.append(label04);
-		label04.update.add(function() {
+		label04.onUpdate.add(() => {
 			if (game.age % rate === 0) {
-				this.rubyOptions.rubyFontSize = counter04 % 5 + 15;
+				label04.rubyOptions.rubyFontSize = counter04 % 5 + 15;
 				counter04++;
-				this.invalidate();
+				label04.invalidate();
 			}
 		}, label04);
 
 		// ルビフォントの指定
 		var label05 = new Label({
 			scene: scene,
-			text:'{"rt":"rubyfont","rb":"ルビフォント"}',
+			text: `{"rt":"rubyfont","rb":"ルビフォント"}`,
 			font: mplusfont,
 			fontSize: 20,
 			width: game.width,
@@ -137,14 +132,14 @@ module.exports = function() {
 		label05.y = y0 + 90;
 		label05.touchable = true;
 		scene.append(label05);
-		label05.update.add(function() {
+		label05.onUpdate.add(() => {
 			if (game.age % rate === 0) {
-				if (this.rubyOptions.rubyFont === bmpfont) {
-					this.rubyOptions.rubyFont = mplusfont;
+				if (label05.rubyOptions.rubyFont === bmpfont) {
+					label05.rubyOptions.rubyFont = mplusfont;
 				} else {
-					this.rubyOptions.rubyFont = bmpfont;
+					label05.rubyOptions.rubyFont = bmpfont;
 				}
-				this.invalidate();
+				label05.invalidate();
 			}
 		}, label05);
 
@@ -152,7 +147,7 @@ module.exports = function() {
 		var y1 = 170;
 		var label11 = new Label({
 			scene: scene,
-			text:'{"rt":"ルビアライン","rb":"ＲｕｂｙＡｌｉｇｎ＝ＳｐａｃｅＡｒｏｕｎｄ"}',
+			text: `{"rt":"ルビアライン","rb":"ＲｕｂｙＡｌｉｇｎ＝ＳｐａｃｅＡｒｏｕｎｄ"}`,
 			font: mplusfont,
 			fontSize: 20,
 			width: game.width
@@ -163,7 +158,7 @@ module.exports = function() {
 		// ルビ位置の調整 Center
 		var label12 = new Label({
 			scene: scene,
-			text:'{"rt":"ルビアライン","rb":"ＲｕｂｙＡｌｉｇｎ＝Ｃｅｎｔｅｒ"}',
+			text: `{"rt":"ルビアライン","rb":"ＲｕｂｙＡｌｉｇｎ＝Ｃｅｎｔｅｒ"}`,
 			font: mplusfont,
 			fontSize: 20,
 			width: game.width,
@@ -183,8 +178,8 @@ module.exports = function() {
 		nlabel.x = 230;
 		nlabel.y = game.height - 20;
 		nlabel.touchable = true;
-		nlabel.pointDown.add(function() {
-			var scene3 = require("mainScene3")();
+		nlabel.onPointDown.add(() => {
+			var scene3 = mainScene3();
 			game.replaceScene(scene3);
 		}, nlabel);
 		scene.append(nlabel);
@@ -193,13 +188,13 @@ module.exports = function() {
 			text: "［フォント切替］",
 			font: mplusfont,
 			fontSize: 20,
-			textAlign: g.TextAlign.Right,
+			textAlign: "right",
 			width: 130
 		});
 		dlabel.x = 100;
 		dlabel.y = game.height - 20;
 		dlabel.touchable = true;
-		dlabel.pointDown.add(function(){
+		dlabel.onPointDown.add(() => {
 			scene.children.forEach((label) => {
 				if (label instanceof Label) {
 					label.font = dfont;
@@ -212,4 +207,4 @@ module.exports = function() {
 
 	});
 	return scene;
-};
+}