+++ sample/src/mainScene3.ts
@@ -1,35 +1,20 @@
-import {Label, RubyAlign} from "@akashic-extension/akashic-label";
-var game = g.game;
+import { Label, RubyAlign } from "@akashic-extension/akashic-label";
+import { mainScene4 } from "./mainScene4";
 
-module.exports = function() {
+export function mainScene3(): g.Scene {
+	var game = g.game;
 	var scene = new g.Scene({
 		game: game,
 		assetIds: ["bmpfont", "bmpfont-glyph", "mplus", "mplus-glyph"]
 	});
-	scene.loaded.add(function() {
+	scene.onLoad.add(() => {
 
 		// グリフデータの生成
-		var mplusGlyph = JSON.parse((<g.TextAsset>scene.assets["mplus-glyph"]).data);
-
+		var mPlusGlyphInfo = JSON.parse(scene.asset.getTextById("mplus-glyph").data);
 		// ビットマップフォント画像とグリフ情報からBitmapFontのインスタンスを生成
 		var mplusfont = new g.BitmapFont({
-			src: scene.assets["mplus"],
-			map: mplusGlyph.map,
-			defaultGlyphWidth: mplusGlyph.width,
-			defaultGlyphHeight: mplusGlyph.height,
-			missingGlyph: mplusGlyph.missingGlyph
-		});
-
-		// グリフデータの生成
-		var glyph = JSON.parse((<g.TextAsset>scene.assets["bmpfont-glyph"]).data);
-
-		// ビットマップフォント画像とグリフ情報からBitmapFontのインスタンスを生成
-		var bmpfont = new g.BitmapFont({
-			src: scene.assets["bmpfont"],
-			map: glyph.map,
-			defaultGlyphWidth: glyph.width,
-			defaultGlyphHeight: glyph.height,
-			missingGlyph: glyph.missingGlyph
+			src: scene.asset.getImageById("mplus"),
+			glyphInfo: mPlusGlyphInfo
 		});
 
 		var dhint: g.DynamicFontHint = {
@@ -38,10 +23,10 @@ module.exports = function() {
 			maxAtlasWidth: 256,
 			maxAtlasHeight: 256,
 			maxAtlasNum: 8
-		}
+		};
 		var dfont = new g.DynamicFont({
 			game: scene.game,
-			fontFamily: g.FontFamily.Monospace,
+			fontFamily: "monospace",
 			size: 40,
 			hint: dhint
 		});
@@ -53,7 +38,7 @@ module.exports = function() {
 			font: mplusfont,
 			fontSize: 31,
 			width: game.width,
-			textAlign: g.TextAlign.Center
+			textAlign: "center"
 		});
 		tlabel0.x = 0;
 		scene.append(tlabel0);
@@ -62,7 +47,7 @@ module.exports = function() {
 		var y0 = 40;
 		var label01 = new Label({
 			scene: scene,
-			text:'ルビの途中でｗｉｄｔｈを超える場合、{"rb": "ｗｉｄｔｈ", "rt": "横幅"}、ルビ内の手前で改行されます',
+			text: `ルビの途中でｗｉｄｔｈを超える場合、{"rb": "ｗｉｄｔｈ", "rt": "横幅"}、ルビ内の手前で改行されます`,
 			font: mplusfont,
 			fontSize: 20,
 			width: game.width
@@ -73,7 +58,7 @@ module.exports = function() {
 		// ルビの幅
 		var label02 = new Label({
 			scene: scene,
-			text:'本文より{"rb": "ルビ", "rt": "とてもながいルビ"}の幅が広い場合、ルビの幅に合わせて余白が設けられます',
+			text: `本文より{"rb": "ルビ", "rt": "とてもながいルビ"}の幅が広い場合、ルビの幅に合わせて余白が設けられます`,
 			font: mplusfont,
 			fontSize: 20,
 			width: game.width
@@ -84,34 +69,34 @@ module.exports = function() {
 		var y1 = 160;
 
 		// ルビ位置の調整とルビ幅の組み合わせ
-		var text11 = '{"rb": "Ａｌｉｇｎ．", "rt": "ルビ", "rubyAlign":' + RubyAlign.Center + '}' +
-					 '{"rb": "Ａｌｉｇｎ．", "rt": "ルビ", "rubyAlign":' + RubyAlign.Center + '}' +
-					 '{"rb": "Ａｒｏｕｎｄ．", "rt": "ルビ", "rubyAlign":' + RubyAlign.SpaceAround + '}' +
-					 '{"rb": "Ａｒｏｕｎｄ．", "rt": "ルビ", "rubyAlign":' + RubyAlign.SpaceAround + '}';
+		var text11 = `{"rb": "Ａｌｉｇｎ．", "rt": "ルビ", "rubyAlign":${RubyAlign.Center}}` +
+					 `{"rb": "Ａｌｉｇｎ．", "rt": "ルビ", "rubyAlign":${RubyAlign.Center}}` +
+					 `{"rb": "Ａｒｏｕｎｄ．", "rt": "ルビ", "rubyAlign":${RubyAlign.SpaceAround}}` +
+					 `{"rb": "Ａｒｏｕｎｄ．", "rt": "ルビ", "rubyAlign":${RubyAlign.SpaceAround}}`;
 		var label11 = new Label({scene: scene, text: text11, font: mplusfont, fontSize: 15, width: game.width, rubyOptions: {rubyFontSize: 10}});
 		label11.y = y1;
 		scene.append(label11);
 
-		var text12 = '{"rt": "Ａｌｉｇｎ．", "rb": "ルビ", "rubyAlign":' + RubyAlign.Center + '}' +
-					 '{"rt": "Ａｌｉｇｎ．", "rb": "ルビ", "rubyAlign":' + RubyAlign.Center + '}' +
-					 '{"rt": "Ａｒｏｕｎｄ．", "rb": "ルビ", "rubyAlign":' + RubyAlign.SpaceAround + '}' +
-					 '{"rt": "Ａｒｏｕｎｄ．", "rb": "ルビ", "rubyAlign":' + RubyAlign.SpaceAround + '}';
+		var text12 = `{"rt": "Ａｌｉｇｎ．", "rb": "ルビ", "rubyAlign":${RubyAlign.Center}}` +
+					 `{"rt": "Ａｌｉｇｎ．", "rb": "ルビ", "rubyAlign":${RubyAlign.Center}}` +
+					 `{"rt": "Ａｒｏｕｎｄ．", "rb": "ルビ", "rubyAlign":${RubyAlign.SpaceAround}}` +
+					 `{"rt": "Ａｒｏｕｎｄ．", "rb": "ルビ", "rubyAlign":${RubyAlign.SpaceAround }}`;
 		var label12 = new Label({scene: scene, text: text12, font: mplusfont, fontSize: 15, width: game.width, rubyOptions: {rubyFontSize: 10}});
 		label12.y = y1 + 40;
 		scene.append(label12);
 
-		var text13 = '{"rb": "Ａｌｉｇｎ．", "rt": "ルビ", "rubyAlign":' + RubyAlign.Center + '}' +
-					 '{"rb": "Ａｒｏｕｎｄ．", "rt": "ルビ", "rubyAlign":' + RubyAlign.SpaceAround + '}' +
-					 '{"rb": "Ａｌｉｇｎ．", "rt": "ルビ", "rubyAlign":' + RubyAlign.Center + '}' +
-					 '{"rb": "Ａｒｏｕｎｄ．", "rt": "ルビ", "rubyAlign":' + RubyAlign.SpaceAround + '}';
+		var text13 = `{"rb": "Ａｌｉｇｎ．", "rt": "ルビ", "rubyAlign":${RubyAlign.Center}}` +
+					 `{"rb": "Ａｒｏｕｎｄ．", "rt": "ルビ", "rubyAlign":${RubyAlign.SpaceAround}}` +
+					 `{"rb": "Ａｌｉｇｎ．", "rt": "ルビ", "rubyAlign":${RubyAlign.Center}}` +
+					 `{"rb": "Ａｒｏｕｎｄ．", "rt": "ルビ", "rubyAlign":${RubyAlign.SpaceAround}}`;
 		var label13 = new Label({scene: scene, text: text13, font: mplusfont, fontSize: 15, width: game.width, rubyOptions: {rubyFontSize: 10}});
 		label13.y = y1 + 70;
 		scene.append(label13);
 
-		var text14 = '{"rt": "Ａｌｉｇｎ．", "rb": "ルビ", "rubyAlign":' + RubyAlign.Center + '}' +
-					 '{"rt": "Ａｒｏｕｎｄ．", "rb": "ルビ", "rubyAlign":' + RubyAlign.SpaceAround + '}' +
-					 '{"rt": "Ａｌｉｇｎ．", "rb": "ルビ", "rubyAlign":' + RubyAlign.Center + '}' +
-					 '{"rt": "Ａｒｏｕｎｄ．", "rb": "ルビ", "rubyAlign":' + RubyAlign.SpaceAround + '}';
+		var text14 = `{"rt": "Ａｌｉｇｎ．", "rb": "ルビ", "rubyAlign":${RubyAlign.Center}}` +
+					 `{"rt": "Ａｒｏｕｎｄ．", "rb": "ルビ", "rubyAlign":${RubyAlign.SpaceAround}}` +
+					 `{"rt": "Ａｌｉｇｎ．", "rb": "ルビ", "rubyAlign":${RubyAlign.Center}}` +
+					 `{"rt": "Ａｒｏｕｎｄ．", "rb": "ルビ", "rubyAlign":${RubyAlign.SpaceAround}}`;
 		var label14 = new Label({scene: scene, text: text14, font: mplusfont, fontSize: 15, width: game.width, rubyOptions: {rubyFontSize: 10}});
 		label14.y = y1 + 100;
 		scene.append(label14);
@@ -126,8 +111,8 @@ module.exports = function() {
 		nlabel.x = 230;
 		nlabel.y = game.height - 20;
 		nlabel.touchable = true;
-		nlabel.pointDown.add(function() {
-			var scene3 = require("mainScene4")();
+		nlabel.onPointDown.add(() => {
+			var scene3 = mainScene4();
 			game.replaceScene(scene3);
 		}, nlabel);
 		scene.append(nlabel);
@@ -136,13 +121,13 @@ module.exports = function() {
 			text: "［フォント切替］",
 			font: mplusfont,
 			fontSize: 20,
-			textAlign: g.TextAlign.Right,
+			textAlign: "right",
 			width: 130
 		});
 		dlabel.x = 100;
 		dlabel.y = game.height - 20;
 		dlabel.touchable = true;
-		dlabel.pointDown.add(function(){
+		dlabel.onPointDown.add(() => {
 			scene.children.forEach((label) => {
 				if (label instanceof Label) {
 					label.font = dfont;
@@ -155,4 +140,4 @@ module.exports = function() {
 
 	});
 	return scene;
-};
+}