+++ sample/src/mainScene5.ts
@@ -1,36 +1,27 @@
-import {Label} from "@akashic-extension/akashic-label";
-var game = g.game;
+import { Label } from "@akashic-extension/akashic-label";
+import { mainScene6 } from "./mainScene6";
 
-module.exports = function() {
+export function mainScene5(): g.Scene {
+	var game = g.game;
 	var scene = new g.Scene({
 		game: game,
 		assetIds: ["bmpfont", "bmpfont-glyph", "mplus", "mplus-glyph"]
 	});
 	var rate = game.fps / 2;
-	scene.loaded.add(function() {
+	scene.onLoad.add(() => {
 
 		// グリフデータの生成
-		var mplusGlyph = JSON.parse((<g.TextAsset>scene.assets["mplus-glyph"]).data);
-
+		var mPlusGlyphInfo = JSON.parse(scene.asset.getTextById("mplus-glyph").data);
 		// ビットマップフォント画像とグリフ情報からBitmapFontのインスタンスを生成
 		var mplusfont = new g.BitmapFont({
-			src: scene.assets["mplus"],
-			map: mplusGlyph.map,
-			defaultGlyphWidth: mplusGlyph.width,
-			defaultGlyphHeight: mplusGlyph.height,
-			missingGlyph: mplusGlyph.missingGlyph
+			src: scene.asset.getImageById("mplus"),
+			glyphInfo: mPlusGlyphInfo
 		});
 
-		// グリフデータの生成
-		var glyph = JSON.parse((<g.TextAsset>scene.assets["bmpfont-glyph"]).data);
-
-		// ビットマップフォント画像とグリフ情報からBitmapFontのインスタンスを生成
+		var bmpGlyphInfo = JSON.parse(scene.asset.getTextById("bmpfont-glyph").data);
 		var bmpfont = new g.BitmapFont({
-			src: scene.assets["bmpfont"],
-			map: glyph.map,
-			defaultGlyphWidth: glyph.width,
-			defaultGlyphHeight: glyph.height,
-			missingGlyph: glyph.missingGlyph
+			src: scene.asset.getImageById("bmpfont"),
+			glyphInfo: bmpGlyphInfo
 		});
 
 		var dhint: g.DynamicFontHint = {
@@ -39,10 +30,10 @@ module.exports = function() {
 			maxAtlasWidth: 256,
 			maxAtlasHeight: 256,
 			maxAtlasNum: 8
-		}
+		};
 		var dfont = new g.DynamicFont({
 			game: scene.game,
-			fontFamily: g.FontFamily.Monospace,
+			fontFamily: "monospace",
 			size: 40,
 			hint: dhint
 		});
@@ -54,7 +45,7 @@ module.exports = function() {
 			font: mplusfont,
 			fontSize: 30,
 			width: game.width,
-			textAlign: g.TextAlign.Center
+			textAlign: "center"
 		});
 		tlabel0.x = 0;
 		scene.append(tlabel0);
@@ -65,7 +56,7 @@ module.exports = function() {
 		// ルビの行幅切替
 		var label01 = new Label({
 			scene: scene,
-			text:'ルビのある行と無い行で\rルビによる{"rb": "行間幅設定", "rt": "fixLineGap"}を\r切り替えることができます',
+			text: `ルビのある行と無い行で\rルビによる{"rb": "行間幅設定", "rt": "fixLineGap"}を\r切り替えることができます`,
 			font: mplusfont,
 			fontSize: 20,
 			width: game.width,
@@ -74,7 +65,7 @@ module.exports = function() {
 		});
 		label01.y = y0;
 		label01.touchable = true;
-		label01.update.add(function() {
+		label01.onUpdate.add(() => {
 			if (game.age % rate === 0) {
 				label01.fixLineGap = !label01.fixLineGap;
 				label01.invalidate();
@@ -83,45 +74,45 @@ module.exports = function() {
 		scene.append(label01);
 
 		// ルビ応用
-		var textArray = [].concat('「よろしゅうございます。'.split(/.*?/),
-			'{"rb": "南　　", "rt": "　　　　　　　"}',
+		var textArray = [].concat(`「よろしゅうございます。`.split(/.*?/),
+			`{"rb": "南　　", "rt": "　　　　　　　"}`,
 			30,
-			'{"rb": "南十　", "rt": "　　　　　　　"}',
+			`{"rb": "南十　", "rt": "　　　　　　　"}`,
 			30,
-			'{"rb": "南十字", "rt": "　　　　　　　"}',
+			`{"rb": "南十字", "rt": "　　　　　　　"}`,
 			30,
-			'{"rb": "南十字", "rt": "サ　　　　　　"}',
+			`{"rb": "南十字", "rt": "サ　　　　　　"}`,
 			30,
-			'{"rb": "南十字", "rt": "サウ　　　　　"}',
+			`{"rb": "南十字", "rt": "サウ　　　　　"}`,
 			30,
-			'{"rb": "南十字", "rt": "サウザ　　　　"}',
+			`{"rb": "南十字", "rt": "サウザ　　　　"}`,
 			30,
-			'{"rb": "南十字", "rt": "サウザン　　　"}',
+			`{"rb": "南十字", "rt": "サウザン　　　"}`,
 			30,
-			'{"rb": "南十字", "rt": "サウザンク　　"}',
+			`{"rb": "南十字", "rt": "サウザンク　　"}`,
 			30,
-			'{"rb": "南十字", "rt": "サウザンクロ　"}',
+			`{"rb": "南十字", "rt": "サウザンクロ　"}`,
 			30,
-			'{"rb": "南十字", "rt": "サウザンクロス"}',
-			'へ着きますのは、次の第三時ころになります。」\r'.split(/.*?/),
-			'{"rb": "車", "rt": "　"}',
+			`{"rb": "南十字", "rt": "サウザンクロス"}`,
+			`へ着きますのは、次の第三時ころになります。」\r`.split(/.*?/),
+			`{"rb": "車", "rt": "　"}`,
 			22,
-			'{"rb": "車", "rt": "し"}',
+			`{"rb": "車", "rt": "し"}`,
 			22,
-			'{"rb": "車", "rt": "しゃ"}',
-			'{"rb": "掌", "rt": "　　　"}',
+			`{"rb": "車", "rt": "しゃ"}`,
+			`{"rb": "掌", "rt": "　　　"}`,
 			24,
-			'{"rb": "掌", "rt": "し　　"}',
+			`{"rb": "掌", "rt": "し　　"}`,
 			24,
-			'{"rb": "掌", "rt": "しょ　"}',
+			`{"rb": "掌", "rt": "しょ　"}`,
 			24,
-			'{"rb": "掌", "rt": "しょう"}',
-			'は紙をジョバンニに渡して向うへ行きました。'.split(/.*?/)
+			`{"rb": "掌", "rt": "しょう"}`,
+			`は紙をジョバンニに渡して向うへ行きました。`.split(/.*?/)
 		);
 		var counter = 0;
 		var mlabel = new Label({
 			scene: scene,
-			text:"",
+			text: "",
 			font: mplusfont,
 			fontSize: 20,
 			lineGap: 0,
@@ -131,27 +122,27 @@ module.exports = function() {
 		});
 		mlabel.y = 130;
 		counter = 0;
-		mlabel.textAlign = g.TextAlign.Left;
+		mlabel.textAlign = "left";
 		scene.append(mlabel);
-		mlabel.update.add(function() {
+		mlabel.onUpdate.add(() => {
 			// 初期化と待機
 			if (counter === textArray.length) {
-				this.text = "";
+				mlabel.text = "";
 				counter = -10;
 				return;
 			}
 			if ( counter < textArray.length && game.age % 2 === 0) {
-				if (counter >= 0){
+				if (counter >= 0) {
 					if (typeof textArray[counter] === "number") {
 						var n = textArray[counter];
-						this.text = this.text.substring(0, this.text.length - n);
+						mlabel.text = mlabel.text.substring(0, mlabel.text.length - n);
 						counter += 1;
-						this.text += textArray[counter];
+						mlabel.text += textArray[counter];
 
 					} else {
-						this.text += textArray[counter];
+						mlabel.text += textArray[counter];
 					}
-					this.invalidate();
+					mlabel.invalidate();
 				}
 				counter += 1;
 			}
@@ -162,13 +153,13 @@ module.exports = function() {
 			text: "［フォント切替］",
 			font: mplusfont,
 			fontSize: 20,
-			textAlign: g.TextAlign.Right,
+			textAlign: "right",
 			width: 130
 		});
 		dlabel.x = 100;
 		dlabel.y = game.height - 20;
 		dlabel.touchable = true;
-		dlabel.pointDown.add(function(){
+		dlabel.onPointDown.add(() => {
 			scene.children.forEach((label) => {
 				if (label instanceof Label) {
 					label.font = dfont;
@@ -190,8 +181,8 @@ module.exports = function() {
 		nlabel.x = 230;
 		nlabel.y = game.height - 20;
 		nlabel.touchable = true;
-		nlabel.pointDown.add(function() {
-			var scene3 = require("mainScene6")();
+		nlabel.onPointDown.add(() => {
+			var scene3 = mainScene6();
 			game.replaceScene(scene3);
 		}, nlabel);
 		scene.append(nlabel);
@@ -201,13 +192,13 @@ module.exports = function() {
 			text: "［フォント切替］",
 			font: mplusfont,
 			fontSize: 20,
-			textAlign: g.TextAlign.Right,
+			textAlign: "right",
 			width: 130
 		});
 		dlabel.x = 100;
 		dlabel.y = game.height - 20;
 		dlabel.touchable = true;
-		dlabel.pointDown.add(function(){
+		dlabel.onPointDown.add(() => {
 			scene.children.forEach((child: g.E) => {
 				if (child instanceof Label) {
 					child.font = dfont;
@@ -219,4 +210,4 @@ module.exports = function() {
 		scene.append(dlabel);
 	});
 	return scene;
-};
+}