+++ sample/src/mainScene6.ts
@@ -1,36 +1,22 @@
-import {Label, RubyAlign, Fragment} from "@akashic-extension/akashic-label";
+import { Label, Fragment } from "@akashic-extension/akashic-label";
+import { mainScene } from "./mainScene";
+
 var game = g.game;
 
-module.exports = function() {
+export function mainScene6(): g.Scene {
 	var scene = new g.Scene({
 		game: game,
 		assetIds: ["bmpfont", "bmpfont-glyph", "mplus", "mplus-glyph"]
 	});
 	var rate = game.fps / 2;
-	scene.loaded.add(function() {
+	scene.onLoad.add(() => {
 
 		// グリフデータの生成
-		var mplusGlyph = JSON.parse((<g.TextAsset>scene.assets["mplus-glyph"]).data);
-
+		var mPlusGlyphInfo = JSON.parse(scene.asset.getTextById("mplus-glyph").data);
 		// ビットマップフォント画像とグリフ情報からBitmapFontのインスタンスを生成
 		var mplusfont = new g.BitmapFont({
-			src: scene.assets["mplus"],
-			map: mplusGlyph.map,
-			defaultGlyphWidth: mplusGlyph.width,
-			defaultGlyphHeight: mplusGlyph.height,
-			missingGlyph: mplusGlyph.missingGlyph
-		});
-
-		// グリフデータの生成
-		var glyph = JSON.parse((<g.TextAsset>scene.assets["bmpfont-glyph"]).data);
-
-		// ビットマップフォント画像とグリフ情報からBitmapFontのインスタンスを生成
-		var bmpfont = new g.BitmapFont({
-			src: scene.assets["bmpfont"],
-			map: glyph.map,
-			defaultGlyphWidth: glyph.width,
-			defaultGlyphHeight: glyph.height,
-			missingGlyph: glyph.missingGlyph
+			src: scene.asset.getImageById("mplus"),
+			glyphInfo: mPlusGlyphInfo
 		});
 
 		var dhint: g.DynamicFontHint = {
@@ -39,10 +25,10 @@ module.exports = function() {
 			maxAtlasWidth: 256,
 			maxAtlasHeight: 256,
 			maxAtlasNum: 8
-		}
+		};
 		var dfont = new g.DynamicFont({
 			game: scene.game,
-			fontFamily: g.FontFamily.Monospace,
+			fontFamily: "monospace",
 			size: 40,
 			hint: dhint
 		});
@@ -53,7 +39,7 @@ module.exports = function() {
 			font: mplusfont,
 			fontSize: 30,
 			width: game.width,
-			textAlign: g.TextAlign.Center
+			textAlign: "center"
 		});
 		tlabel0.x = 0;
 		scene.append(tlabel0);
@@ -61,16 +47,16 @@ module.exports = function() {
 		var counter = 0;
 
 		var text = "「これ」と「それ」と「あれ」と「●●」と「これ」と「それ」と「あれ」と「●●」と「これ」と「それ」と「あれ」と「●●」と「これ」と「それ」と「あれ」と「●●」";
-		var sampleRule = function (fragments: Fragment[], index: number) {
+		var sampleRule = (fragments: Fragment[], index: number) => {
 			const ignoreHead = ["」", "』", "】"];
-			const ignoreTail = ["「", "『", "【"]
+			const ignoreTail = ["「", "『", "【"];
 			const headChar = fragments[index];
 			const isHeadCharIgnore = ignoreHead.indexOf(headChar as string) !== -1;
 			if (typeof headChar !== "string") return index;
 			if (isHeadCharIgnore) {
 				return index + 1;
 			} else {
-				const before = fragments[index-1];
+				const before = fragments[index - 1];
 				const isBeforeIgnore = ignoreHead.indexOf(before as string) !== -1;
 				if (!!before && isBeforeIgnore) {
 					return index;
@@ -79,37 +65,37 @@ module.exports = function() {
 				}
 				return index;
 			}
-		}
+		};
 		var lblabel = new Label({
 			scene: scene,
 			text: text,
 			font: mplusfont,
 			fontSize: 15,
-			textAlign: g.TextAlign.Left,
+			textAlign: "left",
 			width: game.width / 4,
 			lineBreak: true,
 			lineBreakRule: sampleRule
 		});
 		lblabel.y = 40;
 		scene.append(lblabel);
-		lblabel.update.add(function() {
+		lblabel.onUpdate.add(() => {
 			if (game.age % rate === 0) {
-				this.width += 5;
-				if (this.width > game.width) this.width = 100;
-				this.invalidate();
+				lblabel.width += 5;
+				if (lblabel.width > game.width) lblabel.width = 100;
+				lblabel.invalidate();
 			}
 		}, lblabel);
 
-		var text = '「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"●●","rb":"●●"}」と' +
-			'「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"●●","rb":"●●"}」と' +
-			'「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"●●","rb":"●●"}」と' +
-			'「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"●●","rb":"●●"}」';
-		var sampleRule = function(fragments: Fragment[], index: number) {
+		var text = `「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"●●","rb":"●●"}」と` +
+			`「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"●●","rb":"●●"}」と` +
+			`「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"●●","rb":"●●"}」と` +
+			`「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"●●","rb":"●●"}」`;
+		var sampleRule = (fragments: Fragment[], index: number) => {
 			const target = fragments[index];
 			if (target === "」") {
 				return index + 1;
 			} else {
-				var before = fragments[index-1];
+				var before = fragments[index - 1];
 				if (!!before && before === "」") {
 					return index;
 				} else if (!!before && before === "「") {
@@ -117,13 +103,13 @@ module.exports = function() {
 				}
 				return index;
 			}
-		}
+		};
 		var lblabel2 = new Label({
 			scene: scene,
 			text: text,
 			font: mplusfont,
 			fontSize: 15,
-			textAlign: g.TextAlign.Left,
+			textAlign: "left",
 			width: game.width / 4,
 			lineBreak: true,
 			widthAutoAdjust: true,
@@ -131,11 +117,11 @@ module.exports = function() {
 		});
 		lblabel2.y = 190;
 		scene.append(lblabel2);
-		lblabel2.update.add(function() {
+		lblabel2.onUpdate.add(() => {
 			if (game.age % rate === 0) {
-				this.width = counter % 20 * 5 + 120;
+				lblabel2.width = counter % 20 * 5 + 120;
 				counter++;
-				this.invalidate();
+				lblabel2.invalidate();
 			}
 		}, lblabel2);
 
@@ -149,8 +135,8 @@ module.exports = function() {
 		nlabel.x = 230;
 		nlabel.y = game.height - 20;
 		nlabel.touchable = true;
-		nlabel.pointDown.add(function() {
-			var scene3 = require("mainScene")();
+		nlabel.onPointDown.add(() => {
+			var scene3 = mainScene();
 			game.replaceScene(scene3);
 		}, nlabel);
 		scene.append(nlabel);
@@ -160,13 +146,13 @@ module.exports = function() {
 			text: "［フォント切替］",
 			font: mplusfont,
 			fontSize: 20,
-			textAlign: g.TextAlign.Right,
+			textAlign: "right",
 			width: 130
 		});
 		dlabel.x = 100;
 		dlabel.y = game.height - 20;
 		dlabel.touchable = true;
-		dlabel.pointDown.add(function(){
+		dlabel.onPointDown.add(() => {
 			scene.children.forEach((label) => {
 				if (label instanceof Label) {
 					label.font = dfont;
@@ -179,4 +165,4 @@ module.exports = function() {
 
 	});
 	return scene;
-};
+}